UltraPunishments_Spanish
Hello, I need the translation of this plugin, there is a part, it would be good to translate the Economy, permisions and shop de Ultra!, my server is in Spanish and I can not be like this, I will try to help in translating the others. I hope you see this, thanks!